% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_choro_maps.R
\docType{data}
\name{state_test_data}
\alias{state_test_data}
\title{NST Choropleth map}
\format{An object of class \code{data.frame} with 500 rows and 2 columns.}
\usage{
state_test_data
}
\arguments{
\item{Data}{Data.frame containing your data

This is the main mapping function, input data and indentify which variable the
data should be split.}
}
\description{
\code{NST_Choro} Use this function to map your data to a CDC-508 compliant choropleth map
}
\keyword{datasets}
